Moved QBusyIndicatorGoodBad into it's own file
